Extension may not available, let application handle it
The extension should be checked only for this library development.

Example:
Developer cannot identify that user use Windows/or/unix based

However, this package have to be added as "require" package for them
but, Windows user won't be able to meet a requirement. so, it won't be able to install the application.

I think in case of extension not available, developer can still decide the behaviour to process as serial instead using this library.